XXH128Hasher: compute consistent ShaderCreateInfo hash between 32 and 64 bits (fix #688)

+++ Graphics/GraphicsTools/src/XXH128Hasher.cpp
@@ -64,7 +64,7 @@ void XXH128State::Update(const ShaderCreateInfo& ShaderCI) noexcept
 {
     ASSERT_SIZEOF64(ShaderCI, 152, "Did you add new members to ShaderCreateInfo? Please handle them here.");
 
-    Update(ShaderCI.SourceLength, // Aka ByteCodeSize
+    Update(static_cast<Uint32>(ShaderCI.SourceLength), // Aka ByteCodeSize
            ShaderCI.EntryPoint,
            ShaderCI.Desc,
            ShaderCI.SourceLanguage,
